Added info doc. Did some more file management.
