--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Tele/experimentalTeleop.java
@@ -1,10 +1,0 @@
-package org.firstinspires.ftc.teamcode.Tele;
-
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-
-import org.firstinspires.ftc.teamcode.Tele.tested.initialize;
-
-@TeleOp
-public class experimentalTeleop {
-
-}
