+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/competition/TeleOp/competitionMain.java
@@ -1,0 +1,4 @@
+package org.firstinspires.ftc.teamcode.competition.TeleOp;
+
+public class competitionMain {
+}
