+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/competition/TeleOp/competition_linSlide.java
@@ -1,0 +1,4 @@
+package org.firstinspires.ftc.teamcode.competition.TeleOp;
+
+public class competition_linSlide {
+}
