--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/competition/competition_carousel.java
@@ -1,4 +1,0 @@
-package org.firstinspires.ftc.teamcode.competition;
-
-public class competition_carousel {
-}
